Fix off-by-one errors
Now confirms much more closely to the approach in sircovid; the data are
assumed to represent the end of a period of exactly one time unit, so there
is no data discarded, but we extrapolate backwards one time.  At the same
time we add in support for a lead-in time (the only irregular time here).

Goes some way towards #5