Squash coverage hole, and with it small bug
Looks like we should be more chill about what we accept in mode
(mrc-3413), but rerunning the particle filter currently does not
work on master if the times passed to particle_filter_data are
integers, and not numeric. This adds a simple test, while converting
to double at the same time.